api/pyproject.toml
[tool.poetry]
name = "dmsapi"
version = "0.1.0"
description = "The backend API implementation for the STAC Data Management Suite"
authors = ["Nathan Clerkx <nathan.clerkx@deltares.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"stac-fastapi.opensearch" = {git = "https://git@github.com/stac-utils/stac-fastapi-elasticsearch-opensearch.git", tag="3.0.0a4", subdirectory="stac_fastapi/opensearch"}
fastapi-sso = "^0.15.0"
authlib = "^1.3.1"
alembic = "^1.13.2"
sqlmodel = "^0.0.21"
psycopg = {extras = ["binary"], version = "^3.2.1"}


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
pytest = "^8.3.1"
pytest-asyncio = "^0.23.8"

[tool.mypy]
disable_error_code = ["import-untyped"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
